+++ apps/ymir-sdl3/src/app/cmdline_opts.hpp
@@ -8,6 +8,7 @@ struct CommandLineOptions {
     std::filesystem::path gameDiscPath;
     std::filesystem::path profilePath;
     bool fullScreen;
+    bool startPaused;
 };
 
 } // namespace app