+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - App: Added option to to toggle fullscreen by double-clicking the display. (#197)
 - App: Added recent games list to File menu. (#196)
+- App: Added command-line option `-P` to force emulator to start paused.
 
 ### Fixes